+++ packages/attest/src/index.ts
@@ -1,3 +1,4 @@
+export {createStorageRecord} from '.artifactMetadata'
 export {AttestOptions, attest} from './attest'
 export {
   AttestProvenanceOptions,
